Fix schema
