+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const getAllContacts = require('./getAllContacts');
+const getById = require('./getById');
+const addNewContact = require('./addNewContact');
+const updateById = require('./updateById');
+const removeById = require('./removeById');
+
+module.exports = {
+  getAllContacts,
+  getById,
+  addNewContact,
+  updateById,
+  removeById,
+};
