+++ controllers/contacts/removeById.js
@@ -1,0 +1,19 @@
+const { removeContact } = require('../../models/contacts');
+
+const { HttpError } = require('../../helpers');
+
+const removeById = async (req, res, next) => {
+  try {
+    const data = await removeContact(req.params.id);
+    if (!data) {
+      throw HttpError(404);
+    }
+    res.json({
+      message: 'Contact deleted',
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = removeById;
