+++ models/contacts.json
@@ -55,14 +55,14 @@
   },
   {
     "id": "10",
-    "name": "Alec Howard",
-    "email": "Donec.elementum@scelerisquescelerisquedui.net",
-    "phone": "(748) 206-2688"
+    "email": "qweqwe@qweqwe",
+    "name": "Serhii",
+    "phone": "(888) 206-2688"
   },
   {
-    "id": "a52e9e3e-edf2-44ba-98ec-e44c18586fb1",
-    "name": "Max",
+    "id": "e4acf17c-8989-4cc8-a081-581f2e658187",
+    "name": "Yana",
     "email": "qweqwe@qweqwe",
-    "phone": "777-77-77"
+    "phone": "333-33-33"
   }
 ]
